# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Grocery Shop API",
        "version": "1.0.0",
        "description": "E-commerce API for testing purposes"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api",
            "description": "Development server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer" },
                    "name": { "type": "string" },
                    "description": { "type": "string" },
                    "price": { "type": "number" },
                    "category": { "type": "string" },
                    "stock": { "type": "integer" },
                    "image": { "type": "string" }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer" },
                    "email": { "type": "string" },
                    "name": { "type": "string" },
                    "role": { "type": "string", "enum": ["user", "admin"] }
                }
            },
...
        }
    },
    "paths": {
        "/auth/register": {
            "post": {
                "summary": "Register new user",
                "tags": ["Authentication"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" },
                                    "name": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "User registered successfully" },
                    "400": { "description": "User already exists" }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "User login",
                "tags": ["Authentication"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Login successful" },
                    "401": { "description": "Invalid credentials" }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "Get all users",
                "description": "Returns a list of all registered users (admin only)",
                "tags": ["Users"],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": { "$ref": "#/components/schemas/User" }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": { "description": "Access token required" },
                    "403": { "description": "Admin access required" }
                }
            },
            "post": {
                "summary": "Create new user",
                "description": "Create a new user (admin only)",
                "tags": ["Users"],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["email", "password", "name"],
                                "properties": {
                                    "email": { "type": "string", "example": "newuser@grocery.com" },
                                    "password": { "type": "string", "example": "Password123!" },
                                    "name": { "type": "string", "example": "New User" },
                                    "role": { "type": "string", "enum": ["user", "admin"], "default": "user" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "User created successfully" },
                    "400": { "description": "Validation error or user already exists" },
                    "401": { "description": "Access token required" },
                    "403": { "description": "Admin access required" }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "summary": "Update user",
                "description": "Update a user by ID (admin only)",
                "tags": ["Users"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "User ID" }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string", "example": "updated@grocery.com" },
                                    "password": { "type": "string", "example": "NewPassword123!" },
                                    "name": { "type": "string", "example": "Updated Name" },
                                    "role": { "type": "string", "enum": ["user", "admin"] }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "User updated successfully" },
                    "400": { "description": "Validation error or email already in use" },
                    "401": { "description": "Access token required" },
                    "403": { "description": "Admin access required" },
                    "404": { "description": "User not found" }
                }
            },
            "delete": {
                "summary": "Delete user",
                "description": "Delete a user by ID (admin only). Cannot delete your own account.",
                "tags": ["Users"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "User ID" }
                ],
                "responses": {
                    "200": { "description": "User deleted successfully" },
                    "400": { "description": "Cannot delete your own account" },
                    "401": { "description": "Access token required" },
                    "403": { "description": "Admin access required" },
                    "404": { "description": "User not found" }
                }
            }
        },
        "/products": {
            "get": {
                "summary": "Get all products",
                "tags": ["Products"],
                "parameters": [
                    { "name": "page", "in": "query", "schema": { "type": "integer" } },
                    { "name": "limit", "in": "query", "schema": { "type": "integer" } },
                    { "name": "category", "in": "query", "schema": { "type": "string" } },
                    { "name": "search", "in": "query", "schema": { "type": "string" } }
                ],
                "responses": {
                    "200": { "description": "List of products" }
                }
            },
            "post": {
                "summary": "Create new product",
                "tags": ["Products"],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/Product" }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Product created" },
                    "403": { "description": "Admin access required" }
                }
            }
        },
        "/products/reset": {
            "post": {
                "summary": "Reset products database to initial 12 products",
                "tags": ["Products"],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Database reset successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": { "type": "string" },
                                        "productsCount": { "type": "integer" },
                                        "nextId": { "type": "integer" }
                                    }
                                }
                            }
                        }
                    },
                    "403": { "description": "Admin access required" }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "summary": "Get product by ID",
                "tags": ["Products"],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
                ],
                "responses": {
                    "200": { "description": "Product details" },
                    "404": { "description": "Product not found" }
                }
            },
            "put": {
                "summary": "Update product",
                "tags": ["Products"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/Product" }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Product updated" },
                    "403": { "description": "Admin access required" },
                    "404": { "description": "Product not found" }
                }
            },
            "delete": {
                "summary": "Delete product",
                "tags": ["Products"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
                ],
                "responses": {
                    "200": { "description": "Product deleted" },
                    "403": { "description": "Admin access required" },
                    "404": { "description": "Product not found" }
                }
            }
        },
        "/cart": {
            "get": {
                "summary": "Get user cart",
                "tags": ["Cart"],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Cart items" }
                }
            }
        },
        "/cart/add": {
            "post": {
                "summary": "Add item to cart",
                "tags": ["Cart"],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "productId": { "type": "integer" },
                                    "quantity": { "type": "integer" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Item added to cart" }
                }
            }
        },
        "/cart/clear": {
            "delete": {
                "summary": "Clear all items from cart",
                "description": "Remove all products from the user's cart",
                "tags": ["Cart"],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Cart cleared successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": { "type": "string", "example": "Cart cleared successfully" },
                                        "cart": { "type": "array", "items": {}, "example": [] }
                                    }
                                }
                            }
                        }
                    },
                    "401": { "description": "Access token required" }
                }
            }
        },
...
    }
}